+++ packages/web-pkg/src/composables/appDefaults/useAppFolderHandling.ts
@@ -1,7 +1,8 @@
 import { Store } from 'vuex'
 import { computed, Ref, ref, unref } from 'vue'
 import { dirname } from 'path'
-import { ClientService, MaybeRef, useAppFileHandling } from 'web-pkg'
+import { ClientService, MaybeRef } from 'web-pkg'
+import { useAppFileHandling } from 'web-pkg/src/composables'
 import { Resource } from 'web-client'
 import { FileContext } from './types'
 import { RouteLocationNormalizedLoaded } from 'vue-router'
