+++ packages/web-pkg/src/index.ts
@@ -1,4 +1,3 @@
-export * from './composables'
 export * from './configuration'
 export * from './helpers'
 export * from './http'
